requirements.txt
beautifulsoup4==4.8.0
certifi==2024.8.30
chardet==3.0.4
colorama==0.4.1
EbookLib==0.17.1
idna==2.8
lxml==5.3.0
numpy==2.1.1
packaging==24.1
pandas==2.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.22.0
selenium==3.141.0
six==1.16.0
soupsieve==2.6
tzdata==2024.2
urllib3==1.25.11
webdriver-manager==4.0.2
